refactor: add metaphlan subpackage
